Fix possible int-overflow on radius-update

+++ BlueMapCommon/src/main/java/de/bluecolored/bluemap/common/rendermanager/MapUpdateTask.java
@@ -100,13 +100,13 @@ public class MapUpdateTask extends CombinedRenderTask<WorldRegionRenderTask> {
 
         Grid regionGrid = world.getRegionGrid();
         Vector2i halfCell = regionGrid.getGridSize().div(2);
-        int increasedRadiusSquared = (int) Math.pow(radius + Math.ceil(halfCell.length()), 2);
+        long increasedRadiusSquared = (long) Math.pow(radius + Math.ceil(halfCell.length()), 2);
 
         for (Vector2i region : world.listRegions()) {
             Vector2i min = regionGrid.getCellMin(region);
             Vector2i regionCenter = min.add(halfCell);
 
-            if (regionCenter.distanceSquared(center) <= increasedRadiusSquared)
+            if (regionCenter.toLong().distanceSquared(center.toLong()) <= increasedRadiusSquared)
                 regions.add(region);
         }
